Various bug fixes

+++ ribohmm/core/seq.py
@@ -77,18 +77,18 @@ class RnaSequence(object):
         # cdef np.ndarray start_index
 
         offset = 3
-        M = int(self.S/3) - 1
-        start_index = np.zeros((M,3), dtype=np.uint8)
+        M = int(self.S / 3) - 1
+        start_index = np.zeros((M, 3), dtype=np.uint8)
         for f in range(3):  # TODO Python 2/3 compatibility
-            for s in range(f,3*M+f,3):  # TODO Python 2/3
+            for s in range(f, 3 * M + f, 3):  # TODO Python 2/3
                 try:
-                    start_index[s/3,f] = STARTS[self.sequence[s+offset:s+offset+3]]
+                    start_index[int(s/3), f] = STARTS[self.sequence[s + offset:s + offset + 3]]
                 except KeyError:
                     pass
-                for k in [3,6,9,12]:
+                for k in [3, 6, 9, 12]:
                     try:
-                        STOPS[self.sequence[s+offset+k:s+offset+3+k]]
-                        start_index[s/3,f] = 0
+                        STOPS[self.sequence[s + offset + k:s + offset + 3 + k]]
+                        start_index[int(s/3), f] = 0
                     except KeyError:
                         pass
 
@@ -104,12 +104,12 @@ class RnaSequence(object):
         # cdef np.ndarray stop_index
 
         offset = 6
-        M = int(self.S/3) - 1
+        M = int(self.S / 3) - 1
         stop_index = np.zeros((M, 3), dtype=np.uint8)
         for f in range(3):
-            for s in range(f, 3*M+f, 3):
+            for s in range(f, 3 * M + f, 3):
                 try:
-                    stop_index[s/3,f] = STOPS[self.sequence[s+offset:s+offset+3]]
+                    stop_index[int(s/3), f] = STOPS[self.sequence[s + offset:s + offset + 3]]
                 except KeyError:
                     pass
 
@@ -132,7 +132,7 @@ class RnaSequence(object):
             raise
         for f in range(3):
             for s in range(2, int((self.S-f-4-offset) / 3)):
-                score[s,f] = RnaSequence.pwm_score(self.sequence[3*s+offset+f-9:3*s+offset+f+4])
+                score[s, f] = RnaSequence.pwm_score(self.sequence[3*s+offset+f-9:3*s+offset+f+4])
 
         return score  # np.array[*, *]
